Feature/16 extra api

+++ src/main/java/com/example/startlight/member/dto/MemberDto.java
@@ -15,7 +15,7 @@ public class MemberDto {
 
     private String st_nickname;
 
-    private Long memory_num;
+    private Integer memory_num;
 
     public static MemberDto toDto(Member member) {
         return MemberDto.builder()
