+++ src/main/java/com/example/startlight/starList/dto/StarListReqDto.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @Getter
 @Builder
 public class StarListReqDto {
-    private Long x_star;
-    private Long y_star;
+    private Integer x_star;
+    private Integer y_star;
 }
